Add files via upload
As seen in the graph, reservations cancelled when Average Daily rate(ADR) is higher than when it is not cancelled. It clearly proves all the above analysis, that the higher price leads to higher cancellation.

Suggestions
1.	Cancellation rises as the prices does, in order to prevent cancellations, hotels would work on their pricing strategy and try to low the rates for specific hotels as per locations. They can also provide some discounts to retain the consumers.
2.	As the ration of cancellation and not cancellation of resort is higher than the city Hotel, so the hotels should provide the reasonable prices on weekends and on holidays to increase the sales.
3.	In the month of January, hotels can start campaigns or marketing with a reasonable amount to increase their revenue.
4.	They can also increase the quality of Hotels and ambience , services in the Portugal Hotels.

Hypothesis:
1.	More cancellations occurred when prices are higher.
2.	When there is a longer waiting list, customer tend to cancel more frequently.
3.	The majority of client coming from offline travel agents to make their reservations.
Analysis and findings:
 
The Accompany bar graph shows the percentage of reservation that are cancelled and those that are not. It is obvious that there are still a significant number of reservations that have been cancelled. There are 37% of clients who cancelled their reservation, which has a significant impact on the hotels earnings.

 
In comparison to hotel resort, City resort is having more booking, Its possible that resorts are more expensive than the city Hotels.

 
The line graph above shows that, on certain days, the average daily rate for a city hotel is less than that of resort hotel, and on other days its even less. It goes without saying that weekends and holidays may see price rises  in resort hotel rates.

 
We have developed the grouped bar graph to analyze the months with the highest and lowest reservations levels according to reservation status. As we can see for both reservations cancelled and not cancelled have highest no. of records in the month of August, where’s as January is with the most cancelled reservations.

 
This bar graphs proves that cancellations are most common when the prices are higher and are least common when they are lowest. Therefore, the cost of accomaodation is solely responsible for cancellations.

 
Now let see which country has highest cancellations. Portugal is with 70% of cancellati0ns